Restructure the app to correct all offenses
